requirements.txt
pandas==1.5.3
scikit-learn==1.2.2
kafka-python==2.0.2
websockets==11.0.2
pickle-mixin==1.0.2